+++ crates/fluvio-smartengine/src/smartmodule/filter_map.rs
@@ -1,20 +1,37 @@
 use std::convert::TryFrom;
 use anyhow::Result;
-use fluvio_spu_schema::server::stream_fetch::ARRAY_MAP_WASM_API;
-use wasmtime::TypedFunc;
+use wasmtime::{AsContextMut, Trap, TypedFunc};
 
 use dataplane::smartmodule::{SmartModuleInput, SmartModuleOutput, SmartModuleInternalError};
-use crate::smartmodule::{
-    SmartEngine, SmartModuleWithEngine, SmartModuleContext, SmartModuleInstance,
-    SmartModuleExtraParams,
+use crate::{
+    WasmSlice,
+    smartmodule::{
+        SmartEngine, SmartModuleWithEngine, SmartModuleContext, SmartModuleInstance,
+        SmartModuleExtraParams,
+    },
 };
 
 const FILTER_MAP_FN_NAME: &str = "filter_map";
-type FilterMapFn = TypedFunc<(i32, i32), i32>;
+type OldFilterMapFn = TypedFunc<(i32, i32), i32>;
+type FilterMapFn = TypedFunc<(i32, i32, u32), i32>;
 
 pub struct SmartModuleFilterMap {
     base: SmartModuleContext,
-    filter_map_fn: FilterMapFn,
+    filter_map_fn: FilterMapFnKind,
+}
+
+enum FilterMapFnKind {
+    Old(OldFilterMapFn),
+    New(FilterMapFn),
+}
+
+impl FilterMapFnKind {
+    fn call(&self, store: impl AsContextMut, slice: WasmSlice) -> Result<i32, Trap> {
+        match self {
+            Self::Old(filter_fn) => filter_fn.call(store, (slice.0, slice.1)),
+            Self::New(filter_fn) => filter_fn.call(store, slice),
+        }
+    }
 }
 
 impl SmartModuleFilterMap {
@@ -22,11 +39,20 @@ impl SmartModuleFilterMap {
         engine: &SmartEngine,
         module: &SmartModuleWithEngine,
         params: SmartModuleExtraParams,
+        version: i16,
     ) -> Result<Self> {
-        let mut base = SmartModuleContext::new(engine, module, params)?;
-        let filter_map_fn: FilterMapFn = base
+        let mut base = SmartModuleContext::new(engine, module, params, version)?;
+        let filter_map_fn = if let Ok(fmap_fn) = base
             .instance
-            .get_typed_func(&mut base.store, FILTER_MAP_FN_NAME)?;
+            .get_typed_func(&mut base.store, FILTER_MAP_FN_NAME)
+        {
+            FilterMapFnKind::New(fmap_fn)
+        } else {
+            let fmap_fn: OldFilterMapFn = base
+                .instance
+                .get_typed_func(&mut base.store, FILTER_MAP_FN_NAME)?;
+            FilterMapFnKind::Old(fmap_fn)
+        };
 
         Ok(Self {
             base,
@@ -37,7 +63,7 @@ impl SmartModuleFilterMap {
 
 impl SmartModuleInstance for SmartModuleFilterMap {
     fn process(&mut self, input: SmartModuleInput) -> Result<SmartModuleOutput> {
-        let slice = self.base.write_input(&input, ARRAY_MAP_WASM_API)?;
+        let slice = self.base.write_input(&input)?;
         let map_output = self.filter_map_fn.call(&mut self.base.store, slice)?;
 
         if map_output < 0 {
@@ -46,7 +72,7 @@ impl SmartModuleInstance for SmartModuleFilterMap {
             return Err(internal_error.into());
         }
 
-        let output: SmartModuleOutput = self.base.read_output(ARRAY_MAP_WASM_API)?;
+        let output: SmartModuleOutput = self.base.read_output()?;
         Ok(output)
     }