+++ crates/fluvio-smartengine/src/smartmodule/join.rs
@@ -2,21 +2,38 @@ use std::convert::TryFrom;
 
 use anyhow::Result;
 use tracing::{debug, instrument};
-use wasmtime::TypedFunc;
+use wasmtime::{AsContextMut, Trap, TypedFunc};
 
-use fluvio_spu_schema::server::stream_fetch::SMART_MODULE_API;
 use dataplane::smartmodule::{SmartModuleInput, SmartModuleOutput, SmartModuleInternalError};
-use crate::smartmodule::{
-    SmartEngine, SmartModuleWithEngine, SmartModuleContext, SmartModuleInstance,
-    SmartModuleExtraParams,
+use crate::{
+    WasmSlice,
+    smartmodule::{
+        SmartEngine, SmartModuleWithEngine, SmartModuleContext, SmartModuleInstance,
+        SmartModuleExtraParams,
+    },
 };
 
 const JOIN_FN_NAME: &str = "join";
-type JoinFn = TypedFunc<(i32, i32), i32>;
+type OldJoinFn = TypedFunc<(i32, i32), i32>;
+type JoinFn = TypedFunc<(i32, i32, u32), i32>;
 
 pub struct SmartModuleJoin {
     base: SmartModuleContext,
-    join_fn: JoinFn,
+    join_fn: JoinFnKind,
+}
+
+pub enum JoinFnKind {
+    Old(OldJoinFn),
+    New(JoinFn),
+}
+
+impl JoinFnKind {
+    fn call(&self, store: impl AsContextMut, slice: WasmSlice) -> Result<i32, Trap> {
+        match self {
+            Self::Old(join_fn) => join_fn.call(store, (slice.0, slice.1)),
+            Self::New(join_fn) => join_fn.call(store, slice),
+        }
+    }
 }
 
 impl SmartModuleJoin {
@@ -24,12 +41,18 @@ impl SmartModuleJoin {
         engine: &SmartEngine,
         module: &SmartModuleWithEngine,
         params: SmartModuleExtraParams,
+        version: i16,
     ) -> Result<Self> {
-        let mut base = SmartModuleContext::new(engine, module, params)?;
-        let join_fn: JoinFn = base
-            .instance
-            .get_typed_func(&mut base.store, JOIN_FN_NAME)?;
-
+        let mut base = SmartModuleContext::new(engine, module, params, version)?;
+        let join_fn =
+            if let Ok(join_fn) = base.instance.get_typed_func(&mut base.store, JOIN_FN_NAME) {
+                JoinFnKind::New(join_fn)
+            } else {
+                let join_fn = base
+                    .instance
+                    .get_typed_func(&mut base.store, JOIN_FN_NAME)?;
+                JoinFnKind::Old(join_fn)
+            };
         Ok(Self { base, join_fn })
     }
 }
@@ -37,7 +60,7 @@ impl SmartModuleJoin {
 impl SmartModuleInstance for SmartModuleJoin {
     #[instrument(skip(self, input), name = "Join")]
     fn process(&mut self, input: SmartModuleInput) -> Result<SmartModuleOutput> {
-        let slice = self.base.write_input(&input, SMART_MODULE_API)?;
+        let slice = self.base.write_input(&input)?;
         debug!(len = slice.1, "WASM SLICE");
         let map_output = self.join_fn.call(&mut self.base.store, slice)?;
 
@@ -47,7 +70,7 @@ impl SmartModuleInstance for SmartModuleJoin {
             return Err(internal_error.into());
         }
 
-        let output: SmartModuleOutput = self.base.read_output(SMART_MODULE_API)?;
+        let output: SmartModuleOutput = self.base.read_output()?;
         Ok(output)
     }