+++ crates/fluvio-smartengine/src/smartmodule/mod.rs
@@ -31,11 +31,13 @@ pub mod join;
 pub mod file_batch;
 pub mod join_stream;
 
-pub type WasmSlice = (i32, i32);
+pub type WasmSlice = (i32, i32, u32);
 #[cfg(feature = "smartmodule")]
 use fluvio_controlplane_metadata::smartmodule::{SmartModuleSpec};
 
 use self::join_stream::SmartModuleJoinStream;
+
+const DEFAULT_SMARTENGINE_VERSION: i16 = 16;
 
 #[derive(Default, Clone)]
 pub struct SmartEngine(pub(crate) Engine);
@@ -75,25 +77,35 @@ impl SmartEngine {
     pub fn create_module_from_payload(
         self,
         smart_payload: LegacySmartModulePayload,
+        maybe_version: Option<i16>,
     ) -> Result<Box<dyn SmartModuleInstance>> {
+        let version = maybe_version.unwrap_or(DEFAULT_SMARTENGINE_VERSION);
         let smartmodule = self.create_module_from_binary(&smart_payload.wasm.get_raw()?)?;
         let smartmodule_instance: Box<dyn SmartModuleInstance> = match &smart_payload.kind {
-            SmartModuleKind::Filter => Box::new(smartmodule.create_filter(smart_payload.params)?),
+            SmartModuleKind::Filter => {
+                Box::new(smartmodule.create_filter(smart_payload.params, version)?)
+            }
             SmartModuleKind::FilterMap => {
-                Box::new(smartmodule.create_filter_map(smart_payload.params)?)
+                Box::new(smartmodule.create_filter_map(smart_payload.params, version)?)
             }
-            SmartModuleKind::Map => Box::new(smartmodule.create_map(smart_payload.params)?),
+            SmartModuleKind::Map => {
+                Box::new(smartmodule.create_map(smart_payload.params, version)?)
+            }
             SmartModuleKind::ArrayMap => {
-                Box::new(smartmodule.create_array_map(smart_payload.params)?)
+                Box::new(smartmodule.create_array_map(smart_payload.params, version)?)
             }
-            SmartModuleKind::Join(_) => Box::new(smartmodule.create_join(smart_payload.params)?),
+            SmartModuleKind::Join(_) => {
+                Box::new(smartmodule.create_join(smart_payload.params, version)?)
+            }
             SmartModuleKind::JoinStream {
                 topic: _,
                 derivedstream: _,
-            } => Box::new(smartmodule.create_join_stream(smart_payload.params)?),
-            SmartModuleKind::Aggregate { accumulator } => {
-                Box::new(smartmodule.create_aggregate(smart_payload.params, accumulator.clone())?)
-            }
+            } => Box::new(smartmodule.create_join_stream(smart_payload.params, version)?),
+            SmartModuleKind::Aggregate { accumulator } => Box::new(smartmodule.create_aggregate(
+                smart_payload.params,
+                accumulator.clone(),
+                version,
+            )?),
         };
         Ok(smartmodule_instance)
     }
@@ -111,33 +123,49 @@ pub struct SmartModuleWithEngine {
 }
 
 impl SmartModuleWithEngine {
-    fn create_filter(&self, params: SmartModuleExtraParams) -> Result<SmartModuleFilter> {
-        let filter = SmartModuleFilter::new(&self.engine, self, params)?;
+    fn create_filter(
+        &self,
+        params: SmartModuleExtraParams,
+        version: i16,
+    ) -> Result<SmartModuleFilter> {
+        let filter = SmartModuleFilter::new(&self.engine, self, params, version)?;
         Ok(filter)
     }
 
-    fn create_map(&self, params: SmartModuleExtraParams) -> Result<SmartModuleMap> {
-        let map = SmartModuleMap::new(&self.engine, self, params)?;
+    fn create_map(&self, params: SmartModuleExtraParams, version: i16) -> Result<SmartModuleMap> {
+        let map = SmartModuleMap::new(&self.engine, self, params, version)?;
         Ok(map)
     }
 
-    fn create_filter_map(&self, params: SmartModuleExtraParams) -> Result<SmartModuleFilterMap> {
-        let filter_map = SmartModuleFilterMap::new(&self.engine, self, params)?;
+    fn create_filter_map(
+        &self,
+        params: SmartModuleExtraParams,
+        version: i16,
+    ) -> Result<SmartModuleFilterMap> {
+        let filter_map = SmartModuleFilterMap::new(&self.engine, self, params, version)?;
         Ok(filter_map)
     }
 
-    fn create_array_map(&self, params: SmartModuleExtraParams) -> Result<SmartModuleArrayMap> {
-        let map = SmartModuleArrayMap::new(&self.engine, self, params)?;
+    fn create_array_map(
+        &self,
+        params: SmartModuleExtraParams,
+        version: i16,
+    ) -> Result<SmartModuleArrayMap> {
+        let map = SmartModuleArrayMap::new(&self.engine, self, params, version)?;
         Ok(map)
     }
 
-    fn create_join(&self, params: SmartModuleExtraParams) -> Result<SmartModuleJoin> {
-        let join = SmartModuleJoin::new(&self.engine, self, params)?;
+    fn create_join(&self, params: SmartModuleExtraParams, version: i16) -> Result<SmartModuleJoin> {
+        let join = SmartModuleJoin::new(&self.engine, self, params, version)?;
         Ok(join)
     }
 
-    fn create_join_stream(&self, params: SmartModuleExtraParams) -> Result<SmartModuleJoinStream> {
-        let join = SmartModuleJoinStream::new(&self.engine, self, params)?;
+    fn create_join_stream(
+        &self,
+        params: SmartModuleExtraParams,
+        version: i16,
+    ) -> Result<SmartModuleJoinStream> {
+        let join = SmartModuleJoinStream::new(&self.engine, self, params, version)?;
         Ok(join)
     }
 
@@ -145,8 +173,10 @@ impl SmartModuleWithEngine {
         &self,
         params: SmartModuleExtraParams,
         accumulator: Vec<u8>,
+        version: i16,
     ) -> Result<SmartModuleAggregate> {
-        let aggregate = SmartModuleAggregate::new(&self.engine, self, params, accumulator)?;
+        let aggregate =
+            SmartModuleAggregate::new(&self.engine, self, params, accumulator, version)?;
         Ok(aggregate)
     }
 }
@@ -156,6 +186,7 @@ pub struct SmartModuleContext {
     instance: Instance,
     records_cb: Arc<RecordsCallBack>,
     params: SmartModuleExtraParams,
+    version: i16,
 }
 
 impl SmartModuleContext {
@@ -163,6 +194,7 @@ impl SmartModuleContext {
         engine: &SmartEngine,
         module: &SmartModuleWithEngine,
         params: SmartModuleExtraParams,
+        version: i16,
     ) -> Result<Self> {
         let mut store = Store::new(&engine.0, ());
         let cb = Arc::new(RecordsCallBack::new());
@@ -188,28 +220,29 @@ impl SmartModuleContext {
             instance,
             records_cb,
             params,
+            version,
         })
     }
 
-    pub fn write_input<E: Encoder>(&mut self, input: &E, version: i16) -> Result<WasmSlice> {
+    pub fn write_input<E: Encoder>(&mut self, input: &E) -> Result<WasmSlice> {
         self.records_cb.clear();
         let mut input_data = Vec::new();
-        input.encode(&mut input_data, version)?;
+        input.encode(&mut input_data, self.version)?;
         debug!(len = input_data.len(), "input data");
         let array_ptr =
             self::memory::copy_memory_to_instance(&mut self.store, &self.instance, &input_data)?;
         let length = input_data.len();
-        Ok((array_ptr as i32, length as i32))
+        Ok((array_ptr as i32, length as i32, self.version as u32))
     }
 
-    pub fn read_output<D: Decoder + Default>(&mut self, version: i16) -> Result<D> {
+    pub fn read_output<D: Decoder + Default>(&mut self) -> Result<D> {
         let bytes = self
             .records_cb
             .get()
             .and_then(|m| m.copy_memory_from(&mut self.store).ok())
             .unwrap_or_default();
         let mut output = D::default();
-        output.decode(&mut std::io::Cursor::new(bytes), version)?;
+        output.decode(&mut std::io::Cursor::new(bytes), self.version)?;
         Ok(output)
     }
 }