+++ crates/fluvio-smartmodule-derive/src/generator/join.rs
@@ -36,7 +36,7 @@ pub fn generate_join_smartmodule(func: &SmartModuleFn, has_params: bool) -> Toke
         mod __system {
             #[no_mangle]
             #[allow(clippy::missing_safety_doc)]
-            pub unsafe fn join(ptr: *mut u8, len: usize) -> i32 {
+            pub unsafe fn join(ptr: *mut u8, len: usize, version: i16) -> i32 {
                 use fluvio_smartmodule::dataplane::smartmodule::{
                     SmartModuleInput, SmartModuleInternalError,
                     SmartModuleRuntimeError, SmartModuleKind, SmartModuleOutput,
@@ -51,19 +51,19 @@ pub fn generate_join_smartmodule(func: &SmartModuleFn, has_params: bool) -> Toke
 
                 let input_data = Vec::from_raw_parts(ptr, len, len);
                 let mut smartmodule_input = SmartModuleInput::default();
-                if let Err(_err) = Decoder::decode(&mut smartmodule_input, &mut std::io::Cursor::new(input_data), fluvio_smartmodule::api_versions::SMART_MODULE_API) {
+                if let Err(_err) = Decoder::decode(&mut smartmodule_input, &mut std::io::Cursor::new(input_data), version) {
                     return SmartModuleInternalError::DecodingBaseInput as i32;
                 }
 
                 let records_input = smartmodule_input.record_data;
                 let mut records: Vec<Record> = vec![];
-                if let Err(_err) = Decoder::decode(&mut records, &mut std::io::Cursor::new(records_input), fluvio_smartmodule::api_versions::SMART_MODULE_API) {
+                if let Err(_err) = Decoder::decode(&mut records, &mut std::io::Cursor::new(records_input), version) {
                     return SmartModuleInternalError::DecodingRecords as i32;
                 };
 
                 let join_last_record_input = smartmodule_input.join_record;
                 let mut join_last_record: Option<Record> = None;
-                if let Err(_err) = Decoder::decode(&mut join_last_record, &mut std::io::Cursor::new(join_last_record_input), fluvio_smartmodule::api_versions::SMART_MODULE_API) {
+                if let Err(_err) = Decoder::decode(&mut join_last_record, &mut std::io::Cursor::new(join_last_record_input), version) {
                     return SmartModuleInternalError::UndefinedRightRecord as i32;
                 };
                 let join_last_record = match join_last_record {
@@ -102,7 +102,7 @@ pub fn generate_join_smartmodule(func: &SmartModuleFn, has_params: bool) -> Toke
 
                 // ENCODING
                 let mut out = vec![];
-                if let Err(_) = Encoder::encode(&mut output, &mut out, fluvio_smartmodule::api_versions::SMART_MODULE_API) {
+                if let Err(_) = Encoder::encode(&mut output, &mut out, version) {
                     return SmartModuleInternalError::EncodingOutput as i32;
                 }